Unreserve paths when discarding a first edition
Previously we only discarded the content when deleting a draft, which
leaves the original path reservations in the Publishing API. This
creates a problem if users begin creating their document in Content
Publisher, but need to start again in Whitehall.

This adds an association to the Edition model to retrieve all its
Revisions, which enables the DeleteDraftService to find and unreserve
all the associated base paths, so they can be used in other apps.

I've also refactored the DeleteDraftService to keep the top-level method
short and at a single level of abstraction.