Add Mail Recipient Interceptor
We don't want users to receive notifications in integration and staging
environments as they could get confused with production notifications. While
Notify allows us to specify certain email addresses to only receive
notifications in integration and staging, if a notification is sent to a
non-allowed email address (for example in future when publishers receive
notifications that their content has been published) then Notify will return a
`BadRequestError` which is not ideal.

To avoid this, we create a Mail Recipient Interceptor to intercept emails that
are sent in integration and staging and route them to a team email address
instead of the original intended recipient.

Note that `mail.to` can be set to either a string or an array, but when
`mail.to` is called it returns an array-like object (see
https://www.rubydoc.info/github/mikel/mail/Mail/Message#to-instance_method).