Remove unnecessary transactions from the services
Before
Previously there was a nested transaction where the parent transaction
didn't perform any database operations and had a confusing rollback
procedure.

If we just removed the nested transaction and did nothing else then
this would not work as expected. The reason for this is that when a
GdsApi exception is thrown the exception would be rescued before the
transaction had a chance to roll back, meaning records could be
updated / created even if a GdsApi error is thrown.

Now
The rescue is outside the only transaction, this means when a GdsApi
exception
is thrown the transaction handles all the roll backs it needs to do,
uninterrupted. Then when this is complete the exception is bubbled up to
the containing method which is then rescued successfully.

Having a single transaction is better due to it being easier to
understand.