Notify editors when a document is published
This adds a PublishMailer and modifies the publish specs to check the
appropriate email is sent when a document is published with or without
review, as well as when subsequent editions are published. In order to
properly check the time printed in the email, the specs modify the
publish time to ensure it's different to when the edition was created.

Enqueuing email deliveries to Sidekiq has the risk of failure due to
interaction with an external system; the failure modes are unclear. A
failure at the point of sending notifications will cause the publish
transaction to rollback, leaving the app out-of-sync with the Publishing
API and preventing the user from continuing to publish. In addition,
since the emails are sent asynchronously, it's also possible that the
emails are delivered before the transaction commits, which is invalid.

The code to send notifications is therefore outside of the transaction
that encapsulates the Publishing API update and the related change to
keep our local DB in-sync, thus avoiding any unwanted rollback behaviour
and ensuring the emails are enqueued for delivery at a valid point.