Refactor how we test async jobs
Previously we used a gem to test jobs enqueued to Sidekiq. The work for
sending async emails makes use of Sidekiq's in-built test helpers, which
also support the same kind of queue inspection provided by this gem; it
therefore makes sense to replace the gem and use the in-built helpers
directly, so that we're testing Sidekiq consistently throughout.