Use ActiveJob as a wrapper over Sidekiq
Previously we used Sidekiq workers directly. This modifies the existing
code to instead implement the ActiveJob framework, using Sidekiq as the
queueing adapter. Since we're going to use ActiveJob for sending emails
it makes sense to use a consistent framework elsewhere if possible.

In the past ActiveJob has been a poor substitute for Sidekiq when it
came to handling failure and retries. But recently ActiveJob has
improved to the extent that it's a viable drop-in replacement for
Sidekiq, provided we disable Sidekiq's native retry feature!

In contrast to Sidekiq, the default behaviour of ActiveJob is not to
retry. Since we want to retry for the majority of errors, we specify
StandardError but override this with a specific case for RecordNotFound
in order to cope with things like sync between environments.

https://github.com/rails/rails/blob/5-1-stable/activejob/CHANGELOG.md