Add flash error to publish workflow
This adds a flash error if a user doesn't select a radio button
(determines whether the document has been reviewed or not) in the
publishing workflow.

Previously, if a user didn't select which status the document is in
(reviewed or needs 2i) from the publishing workflow, then we would
default this value to published but needs 2i.

This is a bad user experiance as we shouldn't determine what state the
document is in just because the user didn't select the appropriate radio
button.