A different take on the Interactor pattern
The Interactor pattern is quite simple, but trying to use it leaves two
issues unresolved: the verbose passing of variables, either with context
or between methods; and the structure of the actions we're trying to
encapsulate, most which have a somewhat hidden lock/release mechanism.

Re-implementing the Interactor pattern locally gives the opportunity to
address some of these issues, but also to go completely off-the-rails!
This is really another example of what the end-result could look like,
which has some good and bad attributes over vanilla Interactors.