Move and rename specs inside /creating_documents directory
The file rename from `choose_format_spec` to
`create_document_format_not_selected_spec` seemed reasonable as it now
follows a pattern of specs where we are testing the journey when no
option is selected (supertype / document type in this instance).

PR where this pattern is introduced:
https://github.com/alphagov/content-publisher/pull/915

Moving the files to workflow also makes sense as these spec are testing
the creating a document workflow.