Merge show specs and stop testing nil users
Previously we tested some aspects of the document summary page as part
of 'workflow', but there is already a spec that tests other aspects of
the content of this page. We were also testing that we cope gracefully
when the editor/creator for a document does not exist.

This merges the specs that check for content on the document summary
page, and removes the edge-case specs for when a editor/creator does not
exist. We now have several scenarios where a user name is required, so
it's more practical to provide a pattern, based on 'title_or_fallback'.