Improve re-init for frontend components in modal
Previously we had to manually interact with the error-summary component,
which is part of govuk-frontend, in order to ensure it was focussed when
the modal displays an error. This replaces the manual code with a call
to re-init all govuk-frontend components on the modal element and its
children - previously this was only possible for the entire page, which
had a real risk of initialising components twice on other elements.

We should still use components with caution in the context of the modal,
as some are still coupled to the outer window e.g. the tabs component
modifies the page URL, which doesn't make sense within a modal.