Fix excess whitespace in the rendered email
Although the rendered emails are text content, they are used to populate
the body of a GOV.UK Notify template, which is in turn rendered with a
subset of markdown. This means we need to tailer our rendered email text
to produce markdown that will render the appropriate text, primarily for
the text part of the multipart emails sent by Notify.