Remove app documentation page
This page hasn't been updated or maintained and I don't think it's
providing anyone with value. I'm removing this as it remains a lone
debug link in the footer and we are moving away from footer links.

It seems a shame removing it as it's still a very good idea to provide
app documentation through the app, but this is only useful if it's
maintained and we don't seem to have an approach for that.