Show error state on radio buttons if not selected
Previously we only displayed an error summary when a radio selection was
not made for publishing, scheduling, or choosing a format, which is
inconsistent with the design system pattern for forms. This modifies the
argument we pass to the radio component to include any error items, for
which it made sense to use the normal requirements data structures, but
without the boilerplate of having an actual checker class.