Change PublishService to pass an Edition
Previously we passed a Document to the PublishService, which resulted in
the associated current_edition becoming stale in the caller. Because the
caller passes an object to the service, it makes sense the service takes
care to ensure the object is up-to-date when it returns to the caller.

This changes the PublishService to pass an Edition instead of a
Document, as well as reload the passed document before it returns. Some
of the private methods in the service have also been refactored to
hopefully make them a little clearer, as with one of the tests, where the
context didn't explicitly relate to the check.