Add modal module to embed YouTube videos
Govspeak supports embedding YouTube videos (only) using the existing
link syntax, but this behaviour is not obvious. This adds a new modal
module to show a single form that generates the appropriate syntax to
show an embedded YouTube video.

Although this workflow doesn't require server interaction, our frontend
components aren't designed to be re-rendered dynamically, which cause
issues when it comes to showing errors e.g. if the user doesn't enter a
title. We don't believe it's practical to try and change our components
to be more dynamic, so the approach taken here is the same as for inline
images, adding the appropriate controller/views on the server side.