Add Content Data URL class
This encapsulates code around links we need to provide to the Content Data app
which, for now, needs to be a link to the data page for a given piece of
content.  I'm unsure if this should really sit in the services directory,
however since we have a precedent for `Edition URL` being there, I will do the
same and we can move at a later date when we've figured out what should live in
services.

We've decided to use document in this class instead of edition so that we can
still provide a link in case where there is a live edition and an edition that
is being worked on within the app.