Display link to content data if published
We use the Content Data URL class to generate a link to the data page for a
piece of content in Content Data. This link only displays if the content has
been published.

We have used an unordered list to display the link along with the 'View
published edition on GOV.UK' link to ensure adequate spacing between them.