Display content data link to shared beta partners
We only want to display a link to Content Data pages if users are current
Content Data beta partners, otherwise they will not be able to view these
pages.  In order to determine whether the link should be displayed we check the
user's `organisation_content_id` against a list of Content Data beta partner
content IDs.  We include the slugs for each organisation in a comment so that
it's easier to understand which organisations are included in the list.