Fix markdown component
This component wasn't working in the component guide because of it's
coupling to an edition instance variable. I've removed this coupling now
and set a conditional for the bits that integrate more into an edition.

Really this whole thing is barely a component anymore as it's becoming
increasingly coupled to an edition and a particular use-case. I've
resisted pulling the trigger on pulling it out of components as I'm not
sure where it should actually go and don't want to remove/refactor a lot
of code on a whim. I'd suggest that it's days are numbered here in it's
current guise though.