Experiment with avoiding context in #call
This replaces the use of context with setting and passing variables in
the call method of the Update and Destroy Interactors. Although passing
variables can be repetitious, in both classes I feel like it helps to
convey the intention of eah method call better, with the compromise that
setting the context must still be done as a final step.

I've tried to limit the #call method so that all the work is done in
private methods, which seems to have helped, as each line reads with a
similar level of clarity and abstraction, comparable to our features.
The context now mainly serves as an abort mechanism, as we no longer
rely on it within or without the interactor classes.