Replace interactor with business_process
Well I didn't see this coming! This reimplement the Update and Destroy
interactors using the business_process gem, which seems to be a better
fit for the previous version of the code, which was largely avoiding the
use of the interactor context; it was getting to the point where the
context introduced more confusion, with little gain in functionality.

The only place which has arguably suffered is that we have to manually
short-circuit in case of failure, although doing this has turned out to
remove the duplication around setting things like the edition and image
revision on the context in both the success and faiure scenarios.