Make interactors more step focused
Following a workshop with Ben we've been able to agree on an approach to
them that is quite focused on declarative steps.

This increases the readability of the call method, but does mean context
is set as a side effect of other methods. Hopefully that isn't too
confusing.

We've kept the delegates to context, they serve the purpose of reducing
context verbosity in the class and defining the available properties
that can be used when operating on the result.

We dropped the initialize with params and user. These were used as a
means to formalise the class interface but this is quite repetitive.