Paste HTML to markdown behind pre release flag
This ensures only users with the `pre_release_features` permission can use the
"Paste HTML to markdown" functionality. In order to achieve this a change has
been made to the Markdown Editor component so that it accepts a
'paste-html-to-govspeak' data attribute which will determine whether the
functionality is made available.