Up the linkables timeout to 3 seconds.
This is very much an attempt to quick fix the common support requests
we're receiving about timeout problems with the application.

This will give us up to 3 seconds before giving up on loading linkables
rather than 1. The most common linkable we see having this problem is
organisations which on April 10th seems to have hit up to 14 seconds 😬.
So doing this will only limit occurrences of this rather than removing
them all.

I chose 3 seconds as I believe this gives us enough time that in most
cases even with a timeout here we'll still return a response before
nginx gives up and returns a 504.