Use consistent updated boolean in interactors
This is an attempt to bring more consistency to the interactors we have
by using the same boolean of `updated` in each of them rather than a
variety of `edition_updated` and `changed`. `edition_updated` doesn't
make sense for all scenarios - namely that it may be set before the
edition is updated.

`updated` is rather generic and should be versatile to use consistently to
represent that an interactor has caused a mutation.

I am a bit worried it might be too generic so I'm in two minds about it.