Change the many assets association on FileAttachment::FileRevision
This changes the approach to managing the assets associated with with a
FileAttachment::FileRevision. The collection approach makes sense for
images where we have a group of different sizes but less so for files
where we will mostly have a single file and occasionally a thumbnail.
Thus rather than having a group this has specific associations.

I was initially tempted to swap the association to be held by the
file_revision model with a link to the Asset, however this wouldn't have
worked because the Asset needs access to the FileRevision and it can't
determine this if it can be associated on different join points.

It's a bit unclear what the future of the ensure_assets method is as I
don't know how that will determine whether a thumnbnail asset should
exist or not for a PDF, but we'll cross that bridge when we come to it.