Update Preview Service to handle uploading file attachments
This modifies `#upload_images` to be less tied to images and also handle
uploading attachments, as Asset Manager doesn't see any different between them.
At this point in time a File Attachment Revision will only have a file asset
(and we have decided there will be no collection of asset variants).  However
we expect at some point in the future to introduce a thumbnail asset which will
need to be dealt with.