Backtrack on specific asset associations
This essentially reverts changes made in https://github.com/alphagov/content-publisher/pull/975

There we updated a FileAttachment::FileRevision to have a file_asset
association on the logic that we don't anticipate a variety of assets
and generally just a file and a thumbnail. This seemed all well and good
but then meant the logic of file_asset was dripping out into other parts
of the app.

Since that previous change wasn't a clear net improvement it seems more
logical to return to the previous approach as even with any pain points
that is at least consistent with Image::FileRevision.