Rename routes and controller to reflect file attachment
This brings the routes and controllers consistent with the underlying
model and avoids usage of the generic "Attachments". This is because we
are currently only operating on file attachments, with an expectation of
HTML attachments to come.

This does make the UI less consistent with the code where that is
referring to attachment as a singular concept, but hopefully that is of
low concern at this point.