Split adding/updating images/attachments
Previously we used a single 'update' method to handle the addition of
new images and attachments, as well as the substitution of an existing
image or attachment with a new revision. While adding a new image or
attachment happens to be a special case of 'updating', the call to the
'update' method in the context of addition is confusing. This splits the
'add' aspect out of the 'update' method and adds a few pre-checks.