Rename FileRevision to BlobRevision
Previously we used FileRevision to encapsulate the versioning of the
bytes ('blob') associated with an image or a file attachment. This
renames FileRevision to BlobRevision to reduce the reuse of the word
'file' throughout the codebase, as the repetition was confusing.

A separate piece of work will be necessary to make the corresponding
changes in the database, which is preferable to limit the scope of the
breaking change in case a rollback is necessary, and to facilitate the
review of these changes, which touch quite a lot of files.