Publish file attachments to asset manager
File attachments will now be published to asset manager when the
document is published, this will also handle the retirements of assets
and will update the state of the local copys.

PublishAssetService has also been created which moves the code
concerning the publishing of an asset out of the publish_service and
into the publish_asset_service, this seems sensible as the
publish_service was quite clutered with asset related things, i have
also refactored the code so file_attachments and images share as much
code as reasonable.

The pattern images and file_attachments use when retiring assets is,

If:
  * a file attachment is removed from the current edition then it
  should be superseded locally and removed from asset manager.
  * a file attachment is updated from the current edition then the old
  attachment should be superseded and a redirect put in place in asset
  manager to point towards the new attachment
  * the file attachments stays the same then no action is taken