Make api_error booleans consistent in interactors
We had a mix of some places using api_error and some using api_errored.
I had initially planned to use api_errored, which was to remove ambiguity
of whether this was a boolean or a message (as is the case of
api_error), however this isn't a very natural thing to type and is
dubious English.

For consistency I've updated all the places where we were using the term
errored to just be error.