Add file type requirement for attachments
This adds a list of allowed MIME type for a file attachment, based on
the list of extensions in Whitehall. Some of the extensions are
represented by the same MIME type (e.g. ps, eps). We've also removed
some of the allowed files after analysing their use in Whitehall.

  chm - UNUSED - 398a6dfaa094541e48561a2f73bead51e6ce5097
  diff - UNUSED - 398a6dfaa094541e48561a2f73bead51e6ce5097
  dot - UNUSED - 188dc1c65845e5fa2a28a1a8cfb1e3d782921005
  gml - UNUSED - 7b9b4835fc4e9bf2f031caa59bd4933ec3726e5b
  ics - NOT USED RECENTLY - a6b350d2cccc4059c235a975c1bdc8e911321520
  kml - UNUSED - 5219dd96f0585b299f9c50c3c95390d6c8f94234
  rdf - NOT USED RECENTLY - 0d24b52cca263ec6def425a1e9005f5422759deb
  ris - UNUSED - d4e45de654a42bc328f46e3731c62234ba44692e
  sch - UNUSED - 398a6dfaa094541e48561a2f73bead51e6ce5097
  vcf - UNUSED - 234eb8f010927dfdfb2f8173508e9f0a649d93c7
  wsdl - UNUSED - 398a6dfaa094541e48561a2f73bead51e6ce5097
  xslt - UNUSED - 398a6dfaa094541e48561a2f73bead51e6ce5097

We also aim to support ZIP files in future work. Although some of the
now-excluded type could be present in ZIP files, we don't believe
there's enough evidence to support them in the first iteration of the
file attachment feature; we can add them later if necessary.

Files with no extension are also rejected unless the mimetype can be
inferred from the file itself or by the mimetype provided with the file.
For example, text files generally can only be recognised by their
extension / provided mimetype and so if these are not provided the
default type returned by Marcel::MimeType is "application/octet-stream"
which isn't an allowed format. This doesn't seem to be too much of an
issue as only a few files lack a file extension in Asset Manager (where
they're stored for use on gov.uk).