Install pdf-reader gem
This is with the intention to use it to count the number of pages in a PDF as
Active Storage doesn't provide this functionality.  We've chosen to use
pdf-reader as it's widely used and regularly maintained. It's also currently
used by Whitehall to count pages.

We've tested the speed that pdf-reader can count pages for PDFs of different
sizes:

- 227 page PDF
```
puts Benchmark.measure { PDF::Reader.new(file).page_count }
 => 0.010829   0.001643   0.012472 (  0.013456)
```

- 5073 page PDF (currently the PDF with the largest number of pages in Whitehall)
```
puts Benchmark.measure { PDF::Reader.new(file).page_count }
 => 1.210573   0.400807   1.611380 (  1.614067)
```

While over a second for the largest PDF isn't ideal, we're not expecting PDFs
of that size to be the norm so we're happy to use pdf-reader.