Link up the FileAttachmentUploadChecker with the create action
This now runs the requirements checks everytime a user trys to upload a
file attachment. If the user trys to submit a blank / invalid attachment a
flash error is displayed, and if there is an error concerning the title
then a field error error is also displayed.

Also added `value: params[:title]` to the title component which
backfills the title on failure.