Display attachment links on file attachment page
This makes use of the attachment link component to render an attachment link
'preview' for each uploaded attachment. For now it won't link to anything while
we work out which URL to link to in future work.  We also replace the
attachment fixture we use for tests with a PDF so we can test for PDF-specific
features such as displaying a count of pages.