Add helper for file attachment attributes
There are multiple places where a file attachment is converted into a
hash with the same properties. This creates a helper that can be used to
share these rules and avoid the need to repeat.

The in_app terminology originates from the GovspeakDocument set of
classes where in app is used for data used in the app, whereas payload
is used for the Publishing API payload.