Add intelligent preview action for attachments
This adds a 'Preview' action for attachments that redirects to the
external Asset Manager URL if the associated asset is uploaded and
available (scanned for viruses). Further work is needed to:

   - Add a bypass token to the asset URL
   - Use the new preview link elsewhere in the app

I've used an interactor to orchestrate the preview action, but moved
the lower-level upload/pending code into a new PreviewAssetService,
which takes care of DB operations and Asset Manager calls.

An asset on Asset Manager can be in a variety of states (initially,
'unscanned') before it reaches the final 'uploaded' state, which is
when it's (media is) available to request.

If a user tries to preview an attachment that's not yet available (e.g.
because it's just been uploaded), then a 'pending' page is shown that
encourages them to reload (and retry) the preview.