Refactor PreviewService to use PreviewAssetService
Previously the PreviewService and PreviewAssetService both contained
code to manage the upload of assets. This refactors the PreviewService
to delegate to the PreviewAssetService for uploading assets.

Exposing the #upload_asset method, which was previously private, makes
it easier to test the different cases that were previously tied-up in
image and file revisions and their specific assets.

I was a little concerned about exposing the #upload_asset method, as it
relies on the 'asset' data structures for images and attachments having
the same interface. But it turns out we've already made this assumption
in the interface to the AssetManagerService, so it seems just as valid
to re-use that assumption here, not least to minimise duplication.