Add ability to specify margin bottom for metadata component
In using this component I found I needed a margin bottom to it and this
isn't set on this component. There are a few different precedents for
setting margin bottom on components in govuk_publishing_components and
it seems that this approach with a number between 0 and 9 is the most
common.

This component defaults to having no margin bottom to preserve backwards
compatibility.