Add coverage checker for I18n locales
This adds a GOVUKI18nCov class that generates a coverage report of the
I18n keys used after a script (e.g. rspec) has finished running. I've
removed a couple of unused locales that the report has identified as
being unused - the remainder are used but not covered by the tests.

We're unlikely to cover all keys when we have multiple locales, but it
should suffice to strip the locale everywhere and just check if the
remaining key is totally unused from a coverage perspective; in future,
it would be good to check if all used keys are present in all locales!