Add FileAttachmentRevisionUpdater
This is a direct copy of the Image Revision Updater. Currently this doesn't
diverge from the original, so a drawback is that we do have some test
repetition. However when we looked at whether we could have one class that
handled updating both Image and File Attachment revisions, we felt that giving
it a generic name would make it confusing, and even though we have two separate
classes and tests at least it's clear what each of them is for. If in future we
have a new model that has blob and metadata revisions then that would provide a
clearer motivation for some sort of base class that a revision updater for each
model could inherit from, but that's probably overkill right now.