Update File Attachment title
This adds the File Attachment Update Interactor to handle the various updates
that need to occur throughout the app in order to successfully update a File
Attachment title.  At the moment it doesn't handle issues that occur if a title
does not meet requirements; this will be implemented in a future commit.