Use in-app preview path in attachment previews
This updates the file attachment helper to use the new in-app preview
path instead of a placeholder '#'. I've changed all the calls to the
helper to also pass the associated document, which is required to
generate the in-app preview path.

Previously some of the methods used a 'file_attachment' variable to
refer to a 'file_attachment_revision'. This became confusing when I
wrote 'file_attachment.file_attachment', so I've updated the variable to
'attachment_revision', which is a shorthand we use elsewhere.