Rework File Attachment Upload Checker
When updating an existing file attachment's title, we need to check that it
meets the same requirements as the title we set when we upload an attachment.
We do have a File Attachment Upload Checker, however it requires a file in
order to process the checks. When updating an existing file attachment it will
be possible to just edit a title without replacing the file, so we need to
updated the checker to be able to handle just checking the title. This reworks
File Attachment Upload Checker to have two methods that can be called depending
on whether the file attachment is being uploaded or updated. If it's being
updated then we allow a title to be checked, though in future works we expect
the file to also be checked. We also rename the Class to File Attachment
Checker to make it less upload-specific.