Use constants (vs. strings) for instance_doubles
Previously we used instance_double("MyClass") in our tests, which
does nothing (by default) if MyClass doesn't exist. This changes all
usages of instance_double to use the raw constant, which works because
we autoload all classes before the tests run.

Another alternative is to change the configuration of RSpec mocks to
check for class existence on the string form of instance_double, but
this seems unnecessary if we simply change the way we use this feature
instead.