Replace File Attachment file
This allows users to replace a File Attachment file with a new one, provided
the new file has the same filename. In order to do so the new file must be
uploaded and a new blob needs to be created. The File Attachment Upload Service
didn't meet this need as it creates a brand new File Attachment, which we don't
want, as well as new metadata and blob revisions which will be taken care of by
the File Attachment Revision Updater. We just need to create a blob so we
implement the File Attachment Blob Service to do this and access blob data to
be shared with the File Attachment Revision Updater.

We've also found that this service can be used by the File Attachment Upload
Service to create and access data about a blob, meaning that we can reduce some
code duplication. We have extended it to also calculate the number of pages and
handle a new file attachment's requirement to have a unique filename.