Display content data link for all orgs
This code is a little bit redundant as to allow all orgs to view the
link would actually require removing the code around filtering which we can't do
due to it needing to be released at the same time as its communicated.

I've discussed with Tobi and she thinks it should stay behing a flag
until the content is ready to go out communicating the changes.

Trello:
https://trello.com/c/MdKhnoLP/847-link-from-our-app-to-content-data-all-organisations