Fix rollback when creating assets for preview
Previously we did a 'context.fail' when initially uploading an asset for
preview, which caused the wrapping transaction to rollback and the
created asset to be removed, even though it exists in Asset Manager.

This modifies the PreviewInteractor to combine and rename the associated
'check_' methods to better describe what they achieve, which also means
we can short-circuit return and avoid any erreneous rollback.