Separate govspeak options stuff from helpers
Previously we had two helps to generate an hash of attributes for use
with the file attachment component, for in-app and payload use cases.
We now want to show info for how to request an alternative format of an
attachment, but only when rendering the attachments in govspeak (in-app
or payload). This separates the govspeak-specific helpers and moves them
into the associated options classes to cater for the new use case.