Render alternative contact email for attachments
This adds a new OrganisationService to abstract getting the email for
users to request an alternative format of a document. If the document
has no primary organisation, or the organisation has no such email, we
use a default email. We only want to show the email for attachments
rendered via Govspeak, so the helper attributes are unchanged here.