Don't timeout for modal forms that upload files
We originally waited 10 seconds before deciding that a request had timed
out, this can be problematic for file uploads as the 10 seconds includes
the file upload time. This removes the timeout on any forms that are
multipart/form-data on the expectation that they are for uploading
files.

Another change is updating our timeout to 15 seconds, this is so our app
timeout now matches our nginx timeout limit.