Implement initial modal workflow for attachments
Previously clicking 'Attachment' from the insert dropdown on the
markdown toolbar directed the user to a new tab which they could then
see the index for all the documents file attachments (this still happens
in no js experiance).

Now, js enabled, a modal is displayed without having to leave the page.
This follows the same pattern as inserting inline images.