Check for issues before updating file attachment
As updating file attachment involves uploading the file to S3 we should
only do this when we're confident we'll be able to persist the result
otherwise we'll end up with orphaned assets that have no database entry.

This also refactors the ordering of the interactor so that step methods
are before helper ones.