Stop relying on data import for alt email
This changes the OrganisationService to handle the case where the
primary org for an edition is not found in the Publishing API, when
attempting to lookup its alternative format contact email.

Although this is unlikely to be an issue in production, in development
we create a seed user who's primary org may not be in sync with the
development data for the Publishing API.

We should avoid coupling to specific data in remote systems so that our
app continues to work well in a variety of environments, and to minimise
the prerequisites required to begin developing on it.