Refactor rendering/actions into modal workflow
Previously we duplicated some of the render and action code in each
modal module. It now appears three times, which is good evidence it
needs DRY-ing up. This refactors the render and action code into the
existing ModalWorkflow utility to avoid duplication.

Originally it was unclear how portable the code we wrote for inline
images would be, so leaving it in each modal module allowed it to be
easily customised in the future, which a refactoring like this could
make harder. But this flexibility has not been useful, so there's no
strong argument for continuing to duplicate the code between modules.