Make publishing requirement prompt target accurate
Adjusted links found in the warning message box of a draft document such that they bring the user to the required location (e.g. the content edit field of the document) for further editing.

Changed the ID for the govuk content field to refer to revision_field.id instead of just field.id so we avoid conflicts in the use of anchors. Updated related tests that were affected by this.