Refactor actions into helper methods
Previously we had a long and complex 'if' statement to determine which
actions a user can perform on a document. Much of the length of the
statement was due to (duplicated) rendering of elements, passing various
options. This refactors all the rendering into a new 'ActionsHelper' and
an 'external_links' partial, leaving just a complex 'if' statement ;-).