Standardise copy/paste GTM attributes
Previously we enabled tracking for copy/paste events using a particular
'data-gtm-copy-paste-tracking' attribute. In some cases, we also set the
'data-gtm' attribute, which we conventionally use the identify the
element. This ensures 'data-gtm' is populated in all places where we
track copy/paste events, and replaces the (now) redundant value of the
'data-gtm-copy-paste-tracking' attribute with 'true' so it still works.