Pass in only a fields link options for `to_item`
Previously the full set of params would be passed through to an issue
instance and that class would then work out which ones would match the
particular field. This changes it so that only the specific options for
that field are passed in, which moves this logic out of the checker
class.

This has been done to simplify the interface to `to_items` so it's less
coupled to how it's being called.