Retire FileAttachmentUploadService and refactor FileAttachmentBlobService
This removes the FileAttachmentUploadService and instead handles the
concerns of that in a create_initial method on FileAttachment::Revision.
This is to reflect that FileAttachmentUploadService no longer is
responsible for uploading and it's name was becoming redundant.

This also renames FileAttachmentBlobService to
FileAttachmentBlobAttributesService and changes it's interface from
being something that acts like a model into a service that returns a
hash of attributes - hence the name change.